Refactor Seshu to properly use services_* pattern and keep HTTP handler logic only in handlers
