+++ .env.example
@@ -45,3 +45,4 @@ POSTGRES_USER='postgres'
 POSTGRES_PASSWORD='postgres'
 POSTGRES_DB='postgres'
 IS_ACT_LEADER='true'
+DEV_ACT_SSH_KEY='ask_for_key' # SSH key for Act DEV
