+++ functions/gateway/templates/components/navbar.templ
@@ -365,7 +365,7 @@ templ Navbar(userInfo helpers.UserInfo, subnavTabs []string, event types.Event, 
 						>
 							<h3 class="text-xl my-3">Search</h3>
 							<form id="query-search-form" novalidate>
-								<div class="join w-full">
+								<div class="join rounded-full w-full">
 									<div class="w-full">
 										<input
 											id="search-input"
